requirements.txt
openai
firecrawl
openai-agents
python-dotenv
